+++ react.js
@@ -21,12 +21,26 @@ ProvideAtom.prototype.render = function render () {
 function ConnectAtom (props, context) {
   var atom = context.atom
   var map = props.map
-  var render = props.render
+  var render = props.render || props.children
   var data = map ? map(atom.get(), atom.split) : { state: atom.get(), split: atom.split }
   return render(data)
 }
 
 ConnectAtom.contextTypes = { atom: Any }
 
+function connect (map) {
+  return function connectComponent (Component) {
+    return function Connected (props) {
+      return React.createElement(ConnectAtom, {
+        map: map,
+        render: function (mappedProps) {
+          return React.createElement(Component, Object.assign({}, props, mappedProps))
+        }
+      })
+    }
+  }
+}
+
 module.exports.ProvideAtom = ProvideAtom
 module.exports.ConnectAtom = ConnectAtom
+module.exports.connect = connect